+++ packages/ebsr/controller/src/index.js
@@ -5,13 +5,19 @@ import { isResponseCorrect } from './utils';
 
 const log = debug('pie-elements:ebsr:controller');
 
-const prepareChoice = (mode, defaultFeedback) => choice => {
+const prepareChoice = (env, defaultFeedback) => choice => {
   const out = {
     label: choice.label,
     value: choice.value
   };
 
-  if (mode === 'evaluate') {
+  if (env.role === 'instructor' && (env.mode === 'view' || env.mode === 'evaluate')) {
+    out.rationale = choice.rationale;
+  } else {
+    out.rationale = null;
+  }
+
+  if (env.mode === 'evaluate') {
     out.correct = !!choice.correct;
 
     const feedbackType = (choice.feedback && choice.feedback.type) || 'none';
@@ -33,7 +39,7 @@ const parsePart = (part, key, session, env) => {
   );
 
   let choices = part.choices.map(
-    prepareChoice(env.mode, defaultFeedback)
+    prepareChoice(env, defaultFeedback)
   );
 
   if (!part.lockChoiceOrder) {