+++ packages/hotspot/controller/src/index.js
@@ -1,5 +1,7 @@
 import debug from 'debug';
 import isEqual from 'lodash/isEqual';
+import { partialScoring } from '@pie-lib/controller-utils';
+
 import defaults from './defaults';
 
 const log = debug('pie-elements:hotspot:controller');
@@ -10,10 +12,14 @@ const getCorrectResponse = (choices) => choices
   .sort();
 
 const isResponseCorrect = (question, session) => {
-  let correctResponse = getCorrectResponse(question.shapes);
+  const { shapes: { rectangles, polygons } } = question;
+  const choices = [...rectangles, ...polygons];
+  let correctResponse = getCorrectResponse(choices);
 
   if (session.answers.length) {
     return isEqual((session.answers || []).sort(), correctResponse);
+  } else if (!correctResponse.length) {
+    return true;
   }
   return false;
 };
@@ -28,6 +34,7 @@ export function model(question, session, env) {
     maxImageWidth,
     multipleCorrect,
     outlineColor,
+    partialScoring,
     prompt,
     shapes
   } = question;
@@ -43,6 +50,7 @@ export function model(question, session, env) {
       maxImageHeight,
       maxImageWidth,
       multipleCorrect,
+      partialScoring,
       prompt,
       shapes,
       responseCorrect:
@@ -69,17 +77,22 @@ export const createDefaultModel = (model = {}) =>
     })
   });
 
-const getScore = (config, session) => {
+const getScore = (config, session, env) => {
   const { answers } = session;
-  const { partialScoring, shapes } = config;
+  const { shapes: { rectangles, polygons } } = config;
+  const partialScoringEnabled = partialScoring.enabled(config, env);
 
-  if (!partialScoring) {
+  if (!partialScoringEnabled) {
     return isResponseCorrect(config, session) ? 1 : 0;
   }
 
   let correctAnswers = 0;
 
-  shapes.forEach(shape => {
+  console.log('Rectangles: ', rectangles);
+  console.log('polygons: ', polygons);
+
+  const choices = [...rectangles, ...polygons];
+  choices.forEach(shape => {
     const selected = answers.filter(answer => answer.id === shape.id)[0];
     const correctlySelected = shape.correct && selected;
     const correctlyUnselected = !shape.correct && !selected;
@@ -89,16 +102,16 @@ const getScore = (config, session) => {
     }
   });
 
-  const str = (correctAnswers / shapes.length).toFixed(2);
+  const str = (correctAnswers / choices.length).toFixed(2);
   return parseFloat(str);
 };
 
-export function outcome(config, session) {
+export function outcome(config, session, env) {
   return new Promise(resolve => {
     log('outcome...');
 
     if (session.answers) {
-      const score = getScore(config, session);
+      const score = getScore(config, session, env);
       resolve({ score });
     }
   });